requirements.txt
requests==0.11.1
python-telegram-bot==10.1.0
